Extracted shared helpers for the DataFusion runtime and gather drive.
These were duplicated across JoinScan, AggregateScan, and the MPP worker.
`build_runtime_env` centralizes the `work_mem` pool plus disabled disk manager.
`block_on_next` centralizes the gather drive under the cancel/die holdoff. The
AggregateScan teardown collapses to one `datafusion_state` block.
